second commit

+++ src/main/java/com/sae/remisiones/repositorios/primera_escucha_repositorio.java
@@ -8,6 +8,6 @@ import com.sae.remisiones.Entidades.primera_escucha;
 
 @Repository
 @Transactional
-public interface primera_escucha_repositorio extends JpaRepository<primera_escucha, Long> {
+public interface primera_escucha_repositorio extends JpaRepository<primera_escucha, Integer> {
     // Métodos personalizados si es necesario
 }
